type

+++ reverse/reverse.ino
@@ -59,7 +59,7 @@ double getDistance() {
   delayMicroseconds(10);
   digitalWrite(TRIG_PIN, LOW);
   int microsecs = pulseIn(ECHO_PIN, HIGH);
-  float cms_from_wall = microsecs * SPEED_OF_SOUND / 2;
+  double cms_from_wall = microsecs * SPEED_OF_SOUND / 2;
 
   return cms_from_wall;
 }
